Fix g++ carping about overflow in implicit constant conversion for the quiver cookie.
git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@4329 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/quiver.cc
@@ -292,7 +292,7 @@ void player_quiver::_get_fire_order(
 // Save/load
 // ----------------------------------------------------------------------
 
-static const short QUIVER_COOKIE = 0xb015;
+static const short QUIVER_COOKIE = static_cast<short>(0xb015);
 void player_quiver::save(writer& outf) const
 {
     marshallShort(outf, QUIVER_COOKIE);
